# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Österreichische EEG (Energiegemeinschaft) Abrechnungsplattform — REST API",
        "title": "eegabrechnung API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "eegabrechnung Support",
            "email": "support@eeg.at"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8101",
    "basePath": "/api/v1",
    "paths": {
...
        "/public/portal/exchange": {
            "post": {
                "description": "Exchanges a one-time magic link token (from the login email) for a long-lived session token. The session token must be passed as X-Portal-Session header on subsequent portal requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mitgliederportal"
                ],
                "summary": "Exchange magic link token for a portal session",
                "parameters": [
                    {
                        "description": "One-time magic link token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "session_token, member_id, eeg_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Link invalid or expired",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/public/portal/invoices": {
            "get": {
                "description": "Returns all invoices for the authenticated member. Requires a valid X-Portal-Session header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mitgliederportal"
                ],
                "summary": "Get member's invoices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portal session token",
                        "name": "X-Portal-Session",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Invoice"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/public/portal/invoices/{invoiceID}/pdf": {
            "get": {
                "description": "Streams the PDF file for the given invoice. The invoice must belong to the member identified by the X-Portal-Session header. Returns 404 if the PDF has not yet been generated.",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Mitgliederportal"
                ],
                "summary": "Download invoice PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portal session token",
                        "name": "X-Portal-Session",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Invoice UUID",
                        "name": "invoiceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/public/portal/me": {
            "get": {
                "description": "Returns the member record and associated EEG details for the portal session identified by the X-Portal-Session header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mitgliederportal"
                ],
                "summary": "Get authenticated member's profile and EEG info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portal session token",
                        "name": "X-Portal-Session",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "member and eeg fields",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/public/portal/request-link": {
            "post": {
                "description": "Looks up the member by email (and optionally eeg_id) and sends a magic login link by email. If the email is associated with multiple EEGs and no eeg_id is given, returns a list of EEG choices for the user to select. Always returns HTTP 200 to avoid email enumeration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mitgliederportal"
                ],
                "summary": "Request a magic login link for the member portal",
                "parameters": [
                    {
                        "description": "Member email and optional EEG ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "eeg_id": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Invoice": {
            "type": "object",
            "properties": {
                "billing_run_id": {
                    "type": "string"
                },
                "consumption_kwh": {
                    "description": "community kWh consumed",
                    "type": "number"
                },
                "consumption_net_amount": {
                    "description": "Split net amounts — stored separately to enable per-account EA import bookings",
                    "type": "number"
                },
                "consumption_vat_amount": {
                    "type": "number"
                },
                "consumption_vat_pct": {
                    "description": "Split VAT — consumption uses EEG-level rate, generation uses member-specific rate",
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "document_type": {
                    "description": "\"invoice\" or \"credit_note\"",
                    "type": "string"
                },
                "eeg_id": {
                    "type": "string"
                },
                "generation_kwh": {
                    "description": "community kWh generated/fed in",
                    "type": "number"
                },
                "generation_net_amount": {
                    "description": "net Einspeisung (generation credit before VAT)",
                    "type": "number"
                },
                "generation_vat_amount": {
                    "type": "number"
                },
                "generation_vat_pct": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "string"
                },
                "net_amount": {
                    "description": "net amount before VAT",
                    "type": "number"
                },
                "pdf_path": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "sepa_return_at": {
                    "description": "SEPA return (Rücklastschrift) tracking",
                    "type": "string"
                },
                "sepa_return_note": {
                    "type": "string"
                },
                "sepa_return_reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storno_pdf_path": {
                    "description": "set when invoice is formally cancelled (storno document)",
                    "type": "string"
                },
                "total_amount": {
                    "description": "gross saldo incl. VAT (may be negative for producers)",
                    "type": "number"
                },
                "total_kwh": {
                    "description": "consumption kWh (for display/stats)",
                    "type": "number"
                },
                "vat_amount": {
                    "description": "total VAT (consumption + generation)",
                    "type": "number"
                },
                "vat_pct_applied": {
                    "description": "EEG-level VAT rate (consumption side)",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}